Swap new components while removing old components - cleanup
