+++ ui/canvas/NodeTraveseral.ts
@@ -1,18 +1,26 @@
-import {Renderer} from "./Render2"
+import {Renderer, NodeId} from "./Render"
 import {notNull} from "../Utils"
 
+function getRootNodeIDs(graph: any): NodeId[] {
+  return graph.filterNodes((nodeId: NodeId) => { return graph.predecessors(nodeId).length == 0; }).nodes()  
+}
 // Had to make this since couldn't figure out why the built-in dfs was going top / right / left
 export class DFS {
-  visitedNodes: cytoscape.NodeSingular[] = []
-  constructor(rootNodes: cytoscape.NodeCollection) {
-    rootNodes.forEach((node: cytoscape.NodeSingular) => this.continueDFS(node));
+  visitedNodes: NodeId[] = []
+  graph: any | null = null
+  constructor(rootNodes: NodeId[], renderer: Renderer) {
+    this.graph = renderer?.graph
+    if (rootNodes.length == 0) { 
+      rootNodes = getRootNodeIDs(this.graph)  
+    }
+    rootNodes.forEach((nodeID: NodeId) => this.continueDFS(nodeID));
   }
-  continueDFS(currNode: cytoscape.NodeSingular) {
-    this.visitedNodes.push(currNode);
-    currNode.outgoers().forEach((edge: cytoscape.EdgeSingular) => {
-      edge.targets().forEach((child: cytoscape.NodeSingular) => {
-        this.continueDFS(child)
-      });
+  continueDFS(currNodeId: NodeId) {
+    this.visitedNodes.push(currNodeId);
+    this.graph.successors(currNodeId).forEach((childId: any) => {
+      if (this.visitedNodes.indexOf(childId) === -1) {
+        this.continueDFS(childId);
+      }
     });
   }
 };
@@ -20,23 +28,30 @@ export class DFS {
 // Keeps a window of nodes that allows switching the "currently selected node"  
 export class NodeTraverseSelection {
   renderer?: Renderer;
-  curr: cytoscape.NodeSingular;
-  firstParent: cytoscape.NodeSingular | null = null
-  rightSibs: cytoscape.NodeSingular[] = []
-  leftSibs: cytoscape.NodeSingular[] = []
+  curr: NodeId;
+  firstParent: NodeId | null = null
+  rightSibs: NodeId[] = []
+  leftSibs: NodeId[] = []
 
-  constructor(renderer: Renderer, startingNode: cytoscape.NodeSingular) {
+  constructor(renderer: Renderer, startinNodeId: NodeId) {
     this.renderer = renderer;
-    this.curr = startingNode;
-    this.populateFrom(startingNode);
+    this.curr = startinNodeId
+    this.populateFrom(this.curr);
   }
 
-  clear() { this.curr.unselect(); }
+  node() {
+    return this._graph().node(this.curr);
+  }
+
+  clear() { 
+    this.renderer?.onNodeSelect(null)
+  }
 
   moveDown() {
-    // @ts-ignore: targets doesn't exist?
-    const firstChild = this.curr.outgoers()?.first()?.targets()?.first();
-    this.populateFrom(firstChild);
+    const children = this.renderer?.graph.successors(this.curr);
+    if (children && children.length > 0) {
+      this.populateFrom(children[0]);
+    }
   }
 
   moveUp() {
@@ -54,31 +69,21 @@ export class NodeTraverseSelection {
   }
  
   // Populates member variables starting from startingNode to be used for switching which node is selected
-  populateFrom(startingNode: cytoscape.NodeSingular | null) {
-    if (!startingNode || startingNode.length == 0) { return; }
+  populateFrom(startingNodeId: any | null) {
+    if (!startingNodeId || startingNodeId.length == 0) { return; }
     this.clear();
-    this.curr = startingNode;
-    // @ts-ignore: source doesn't exist?
-    this.firstParent = this.curr.incomers()?.first()?.source()?.first();
+    this.curr = startingNodeId;
+    const parents = this._graph().predecessors(this.curr)
+    this.firstParent = (parents && parents.indexOf(this.curr) === -1) ? parents[0] : null;
     // Get the siblings - handle if its a root or if it has actual siblings
-    let sibs = !this.firstParent ? notNull(this.renderer).cy.nodes().roots() : [];
-    if (this.firstParent) {
-      this.firstParent?.outgoers()?.forEach((edge: cytoscape.EdgeSingular) => {
-        edge.targets().forEach((sibOrSelf: cytoscape.NodeSingular) => {
-         sibs.push(sibOrSelf);
-      })});
-    }
+    let sibIds = this.firstParent ? this._graph().successors(this.firstParent) : getRootNodeIDs(this._graph());
     // Separate the siblings between left and right
-    let gettingLeftSibs = true;
-    this.leftSibs = [], this.rightSibs = [];
-    sibs?.forEach((sibOrSelf: cytoscape.NodeSingular) => {
-      if (gettingLeftSibs && sibOrSelf.data().id === this.curr?.data().id) {;
-        gettingLeftSibs = false;
-        return;
-      }
-      let correctSideSibs = gettingLeftSibs ? this.leftSibs : this.rightSibs;
-      correctSideSibs.push(sibOrSelf);
-    });
-    this.curr.select();
+    const splitIdx = sibIds.indexOf(this.curr)
+    this.leftSibs = sibIds.slice(0, splitIdx).reverse()
+    this.rightSibs = sibIds.slice(splitIdx + 1)
+  }
+
+  _graph() {
+    return this.renderer?.graph;
   }
 }
